Render closed tickets. Rename component
