+++ src/views/Chat.js
@@ -11,6 +11,7 @@ class Chat extends React.Component {
         <h2>Welcome to the Chat page!</h2>
         <ChatsList chatsArr={this.props.queuedChats} status={QUEUED} />
         <ChatsList chatsArr={this.props.activeChats} status={ACTIVE} />
+        <ChatsList chatsArr={this.props.closedChats} status={CLOSED} />
       </div>
     );
   }
@@ -20,6 +21,7 @@ const mapStateToProps = state => {
   return {
     queuedChats: state.chats.queuedChats,
     activeChats: state.chats.activeChats,
+    closedChats: state.chats.closedChats,
   };
 };
